feat(zfb): extract useLang() hook; fix SSR-unsafe lang init in LandingLangIsland
Extract shared useLang() hook to components/zfb/use-lang.ts, following the
effect-based SSR-safe pattern from manual-app.tsx. Refactor both islands to
use it, eliminating the duplicated lang state/effect/setLang logic and fixing
LandingLangIsland's previous localStorage read in state initializer (not
SSR-safe). initialLang kept in ManualAppProps (page template still passes it).
