+++ components/zfb/landing-lang-island.tsx
@@ -1,9 +1,8 @@
 'use client';
 
-import { useState, useCallback } from 'preact/hooks';
 import { LanguageToggle } from './language-toggle';
 import type { Lang } from './lang';
-import { DEFAULT_LANG, readPersistedLang, writeLangToStorage, syncLangToUrl } from './lang';
+import { useLang } from './use-lang';
 
 export interface LandingLangIslandProps {
   availableLangs: readonly Lang[];
@@ -13,16 +12,10 @@ export interface LandingLangIslandProps {
  * Island root for the landing-page language toggle. Owns its own lang state
  * so the LanguageToggle is self-contained (not wired to the mega-island, which
  * only mounts on detail pages). Serializable props: only `availableLangs`
- * (string array). Mirrors the ManualApp lang state init pattern.
+ * (string array). Uses useLang() for SSR-safe effect-based localStorage read.
  */
 export default function LandingLangIsland({ availableLangs }: LandingLangIslandProps) {
-  const [lang, setLangState] = useState<Lang>(() => readPersistedLang() ?? DEFAULT_LANG);
-
-  const setLang = useCallback((next: Lang) => {
-    setLangState(next);
-    writeLangToStorage(next);
-    syncLangToUrl(next);
-  }, []);
+  const [lang, setLang] = useLang();
 
   return <LanguageToggle lang={lang} setLang={setLang} availableLangs={availableLangs} />;
 }
